requirements.txt
streamlit>=1.20
python-dotenv>=1.0
pandas>=1.5
langchain
langchain_groq
python-dotenv
